CSER-171 , Mobile site is not displaying pictures

+++ utils/website_functions.php
@@ -112,66 +112,70 @@ function get_website_section_items($db, $section_data) {
             //=======
 
 
-            $image_mobile_website='';
-
-            /*
-
-            require_once 'external_libs/ImageCache.php';
-            $imagecache                         = new ImageCache();
-            $imagecache->cached_image_directory = 'EcomB2B/server_files/cached_images/';
 
 
 
 
-            $image_key = $row['Category Main Image Key'];
-
-            $_size_image_product_webpage = '320_200';
-
-            if ($image_key) {
-
-                $image_format='jpeg';
-                if (file_exists('EcomB2B/server_files/cached_images/'.md5($image_key.'_'.$_size_image_product_webpage.'.'.$image_format).'.'.$image_format)) {
-                    $image_mobile_website = 'server_files/cached_images/'.md5($image_key.'_'.$_size_image_product_webpage.'.'.$image_format).'.'.$image_format;
-                } else {
-                    $image          = get_object('Image', $image_key);
-                    $image_filename = 'EcomB2B/server_files/tmp/'.$image_key.'_'.$_size_image_product_webpage.'.'.$image_format;
-
-                    if (!file_exists($image_filename)  ) {
+            if (is_writable('EcomB2B/server_files/cached_images/')  and is_writable('EcomB2B/server_files/tmp/') ) {
 
 
+                require_once 'external_libs/ImageCache.php';
+                $imagecache                         = new ImageCache();
+                $imagecache->cached_image_directory = 'EcomB2B/server_files/cached_images/';
 
 
-                        $_image= $image->fit_to_canvas(320, 200);
+                $image_key = $row['Category Main Image Key'];
+
+                $_size_image_product_webpage = '320_200';
+
+                if ($image_key) {
+
+                    $image_format = 'jpeg';
+                    if (file_exists('EcomB2B/server_files/cached_images/'.md5($image_key.'_'.$_size_image_product_webpage.'.'.$image_format).'.'.$image_format)) {
+                        $image_mobile_website = 'server_files/cached_images/'.md5($image_key.'_'.$_size_image_product_webpage.'.'.$image_format).'.'.$image_format;
+                    } else {
+                        $image          = get_object('Image', $image_key);
+                        $image_filename = 'EcomB2B/server_files/tmp/'.$image_key.'_'.$_size_image_product_webpage.'.'.$image_format;
+
+                        if (!file_exists($image_filename)) {
 
 
-                        if($image->get('Image File Format')=='png'){
-                            $image->save_image_to_file_as_jpeg('EcomB2B/server_files/tmp', $image_key.'_'.$_size_image_product_webpage,$_image, $image_format);
-                        }else{
-                            $image->save_image_to_file('EcomB2B/server_files/tmp', $image_key.'_'.$_size_image_product_webpage,$_image, $image_format);
+                            $_image = $image->fit_to_canvas(320, 200);
+
+
+                            if ($image->get('Image File Format') == 'png') {
+                                $image->save_image_to_file_as_jpeg('EcomB2B/server_files/tmp', $image_key.'_'.$_size_image_product_webpage, $_image, $image_format);
+                            } else {
+                                $image->save_image_to_file('EcomB2B/server_files/tmp', $image_key.'_'.$_size_image_product_webpage, $_image, $image_format);
+
+                            }
+
 
                         }
-
-
-
-                    }
-
-
 
 
                         $image_mobile_website = $imagecache->cache($image_filename);
 
 
-                    unlink($image_filename);
+                        unlink($image_filename);
+                    }
+
+
+                } else {
+                    $image_mobile_website = '/art/nopic.png';
+
                 }
 
-
-            } else {
+            }else{
                 $image_mobile_website = '/art/nopic.png';
+                $image_key = $row['Category Main Image Key'];
+                if ($image_key) {
 
+                    $image_mobile_website = '/image_root.php?id='.$image_key;
+
+                }
             }
 
-
-*/
 
 
 
@@ -196,6 +200,11 @@ function get_website_section_items($db, $section_data) {
                 
 
             );
+
+
+         //   print_r($categories);
+           // exit;
+
             $category_stack_index = $row['Category Webpage Index Stack'];
             $stack_index++;
         }
